# perf(parser): Specially dispatch hotloops