This took `minimal` from 7.2us to 4.1us when `toml-rs` is 4.9us.

For `medium`, we went from 182us to 160us when `toml-rs` is 127us.

This dropped `Errors::merge` from 7.8% to 4.7% of a sample program's
execution time.